Revert "+deleting favorites"
