+++ client/index.html
@@ -1,16 +1,10 @@
 <html>
   <head>
     <meta charset="UTF-8" />
-    <title>Yeti's Crab Shack & Recipe Finder</title>
+    <title>Scrum Dog Millionaires</title>
   </head>
-  <style>
-    .wong {
-      font-style: italic;
-    }
-  </style>
   <body>
-    <h1>Yeti's Crab Shack & Recipe Finder</h1>
-    <h3>Can't go <span class="wong">wong</span> with our recipes!</h3>
+    <h1>Scrum Dog Millionaires</h1>
     <div id="root"></div>
 
     <!-- script for webpack and webpack-dev-server challenge -->
