+++ server/controllers/userController.js
@@ -17,35 +17,17 @@ userController.addUser = async (req, res, next) => {
   } catch (error) {
     return next(error);
   }
-};
-
-userController.removeFavorite = async (req, res, next) => {
-  console.log('inside userController.removeFavorite');
-  const { image, title, usedIngredients, missedIngredients } = req.body;
-
-  try {
-    const foundItem = await Favorites.findOne({
-      image,
-      title,
-      usedIngredients,
-      missedIngredients,
-    });
-
-    const arr = [];
-    const user = await User.findOne({ username: 'user1' });
-    for (let i = 0; i < user.favorited.length; i++) {
-      if (user.favorited[i].toString() === foundItem._id.toString()) {
-        console.log(i);
-      } else {
-        arr.push(user.favorited[i]);
-      }
-    }
-    await User.updateOne({ username: 'user1' }, { favorited: arr });
-
-    return next();
-  } catch (error) {
-    return next(error);
-  }
+  //   User.create({
+  //     username,
+  //     password,
+  //   })
+  //     .then((createdUser) => {
+  //       console.log(createdUser);
+  //       return next();
+  //     })
+  //     .catch((err) => {
+  //       return next(err);
+  //     });
 };
 
 module.exports = userController;
